Moved TcpServer project to samples/Lab3
