+++ samples/Lab3/NetworkProgramming.Lab3/App.xaml
