+++ samples/Lab3/NetworkProgramming.Lab3/Assets/avalonia-logo.ico
