+++ samples/Lab3/NetworkProgramming.Lab3/Extensions/SocketExtensions.cs
