+++ samples/Lab3/NetworkProgramming.Lab3/Models/ClientModel.cs
@@ -1,0 +1,22 @@
+﻿using System;
+
+namespace NetworkProgramming.Lab2.Models
+{
+	public class ClientModel
+	{
+		public string Id { get; }
+		public string Ip { get; }
+		public int Port { get; }
+
+		public ClientModel(Tuple<int, string> info)
+		{
+			Id = $"Client_{Guid.NewGuid()}";
+			(Port, Ip) = info;
+		}
+
+		public override string ToString()
+		{
+			return $"Id: {Id} | Ip: {Ip} | Port: {Port}";
+		}
+	}
+}
