+++ samples/Lab3/NetworkProgramming.Lab3/Models/InternalMessageModel.cs
