+++ samples/Lab3/NetworkProgramming.Lab3/Models/NetworkInterfaceModel.cs
