+++ samples/Lab3/NetworkProgramming.Lab3/NetworkProgramming.Lab3.csproj
