+++ samples/Lab3/NetworkProgramming.Lab3/ViewLocator.cs
