+++ samples/Lab3/NetworkProgramming.Lab3/ViewModels/ViewModelBase.cs
