+++ samples/Lab3/NetworkProgramming.Lab3/Views/ClientsGrid.xaml
