+++ samples/Lab3/NetworkProgramming.Lab3/Views/MainView.xaml
