+++ samples/Lab3/NetworkProgramming.Lab3/Views/MainWindow.xaml.cs
