+++ samples/Lab3/NetworkProgramming.Lab3/Views/MenuControl.xaml.cs
